requirements.txt
google-generativeai==0.7.2
PyMuPDF==1.24.9
pillow==10.4.0
pandas==2.2.2
tqdm==4.66.5
tenacity==8.5.0
python-dotenv==1.0.1
